requirements.txt
imutils
matplotlib
numpy
hjson
opencv_python
Pillow
scikit_learn
tensorboard
torch_tb_profiler
torch
torchvision
tqdm
efficientnet_pytorch
seaborn
